+++ service/cart/src/main/java/com/bappul/cart/CartApplication.java
@@ -1,0 +1,17 @@
+package com.bappul.cart;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.retry.annotation.EnableRetry;
+
+@EnableRetry
+@EnableFeignClients
+@EnableDiscoveryClient
+@SpringBootApplication
+public class CartApplication {
+  public static void main(String[] args) {
+    SpringApplication.run(CartApplication.class, args);
+  }
+}
