+++ service/cart/src/main/java/com/bappul/cart/client/CatalogClient.java
@@ -1,0 +1,19 @@
+package com.bappul.cart.client;
+
+import com.bappul.cart.client.request.PricingInternalRequest;
+import com.bappul.cart.client.response.PricingInternalResponse;
+import com.bappul.cart.config.OpenFeignConfig;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+
+@FeignClient(
+    name = "catalog-service",
+    configuration = OpenFeignConfig.class
+)
+public interface CatalogClient {
+
+  @PostMapping("/api/v1/menu/price-snapshots")
+  PricingInternalResponse calculate(@RequestBody PricingInternalRequest request);
+
+}
