+++ service/cart/src/main/java/com/bappul/cart/common/exception/GlobalExceptionHandler.java
@@ -1,0 +1,47 @@
+package com.bappul.cart.common.exception;
+
+import exception.ServiceException;
+import java.util.concurrent.atomic.AtomicReference;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindException;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import response.ApiResponse;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+  private final String VALIDATE_ERROR = "VALIDATE_ERROR";
+  private final String SERVER_ERROR = "SERVER_ERROR";
+
+  @ExceptionHandler(ServiceException.class)
+  public ResponseEntity<?> handleResponseException(ServiceException ex){
+    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+        .body(ApiResponse.error(ex.getCode(), ex.getMessage()));
+  }
+
+  @ExceptionHandler(MethodArgumentNotValidException.class)
+  public ResponseEntity<?> methodArgumentNotValidException(MethodArgumentNotValidException ex){
+    AtomicReference<String> errors = new AtomicReference<>("");
+    ex.getBindingResult().getAllErrors().forEach(e -> errors.set(e.getDefaultMessage()));
+    return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+        .body(ApiResponse.badRequest(VALIDATE_ERROR, String.valueOf(errors)));
+  }
+
+  @ExceptionHandler(BindException.class)
+  public ResponseEntity<?> bindException(BindException ex){
+    AtomicReference<String> errors = new AtomicReference<>("");
+    ex.getBindingResult().getAllErrors().forEach(e -> errors.set(e.getDefaultMessage()));
+    return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+        .body(ApiResponse.badRequest(VALIDATE_ERROR, String.valueOf(errors)));
+  }
+
+  @ExceptionHandler(Exception.class)
+  public ResponseEntity<?> serverException(Exception ex){
+    return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+        .body(ApiResponse.serverError(SERVER_ERROR, ex.getMessage()));
+  }
+
+}
