+++ service/cart/src/main/java/com/bappul/cart/domain/entity/Status.java
@@ -1,0 +1,9 @@
+package com.bappul.cart.domain.entity;
+
+public enum Status {
+  ACTIVE,
+  LOCKED,
+  ORDERED,
+  EXPIRED
+  ;
+}
