+++ service/cart/src/main/java/com/bappul/cart/domain/repository/CartItemOptionRepository.java
@@ -1,0 +1,12 @@
+package com.bappul.cart.domain.repository;
+
+import com.bappul.cart.domain.entity.CartItem;
+import com.bappul.cart.domain.entity.CartItemOption;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CartItemOptionRepository extends JpaRepository<CartItemOption, Long> {
+  List<CartItemOption> findAllByCartItem(CartItem cartItem);
+}
