+++ service/cart/src/main/java/com/bappul/cart/domain/repository/CartItemRepository.java
@@ -1,0 +1,12 @@
+package com.bappul.cart.domain.repository;
+
+import com.bappul.cart.domain.entity.Cart;
+import com.bappul.cart.domain.entity.CartItem;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+  List<CartItem>  findAllByCart(Cart cart);
+}
