+++ service/cart/src/main/java/com/bappul/cart/domain/repository/CartRepository.java
@@ -1,0 +1,14 @@
+package com.bappul.cart.domain.repository;
+
+import com.bappul.cart.domain.entity.Cart;
+import com.bappul.cart.domain.entity.Status;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CartRepository extends JpaRepository<Cart, Long> {
+  Optional<Cart> findByUserIdAndStoreIdAndStatus(Long userId, Long storeId, Status status);
+  Optional<Cart> findByUserId(Long userId);
+  Boolean existsByUserId(Long userId);
+}
