+++ service/cart/src/main/java/com/bappul/cart/web/v1/request/CartItemRequest.java
@@ -1,0 +1,11 @@
+package com.bappul.cart.web.v1.request;
+
+import java.util.List;
+import lombok.Getter;
+
+@Getter
+public class CartItemRequest {
+  Long menuId;
+  List<Long> optionValueIds;
+  int quantity;
+}
