+++ service/cart/src/main/java/com/bappul/cart/web/v1/request/CartRequest.java
@@ -1,0 +1,17 @@
+package com.bappul.cart.web.v1.request;
+
+import java.util.List;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class CartRequest {
+  Long storeId;
+  List<CartItemRequest> items;
+}
