--- service/cart/src/main/java/com/bappul/delivery/cart/client/CatalogClient.java
@@ -1,19 +1,0 @@
-package com.bappul.delivery.cart.client;
-
-import com.bappul.delivery.cart.client.request.PricingInternalRequest;
-import com.bappul.delivery.cart.client.response.PricingInternalResponse;
-import com.bappul.delivery.cart.config.OpenFeignConfig;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-
-@FeignClient(
-    name = "catalog-service",
-    configuration = OpenFeignConfig.class
-)
-public interface CatalogClient {
-
-  @PostMapping("/api/v1/menu/price-snapshots")
-  PricingInternalResponse calculate(@RequestBody PricingInternalRequest request);
-
-}
