--- service/cart/src/main/java/com/bappul/delivery/cart/client/request/PricingInternalRequest.java
@@ -1,18 +1,0 @@
-package com.bappul.delivery.cart.client.request;
-
-import com.bappul.delivery.cart.web.v1.request.CartItemRequest;
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class PricingInternalRequest {
-  Long storeId;
-  List<CartItemRequest> items;
-}
