--- service/cart/src/main/java/com/bappul/delivery/cart/client/response/OptionPerPrice.java
@@ -1,18 +1,0 @@
-package com.bappul.delivery.cart.client.response;
-
-import java.math.BigDecimal;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class OptionPerPrice {
-  Long optionValueId;
-  String optionName;
-  BigDecimal optionPrice;
-}
