--- service/cart/src/main/java/com/bappul/delivery/cart/domain/entity/Status.java
@@ -1,9 +1,0 @@
-package com.bappul.delivery.cart.domain.entity;
-
-public enum Status {
-  ACTIVE,
-  LOCKED,
-  ORDERED,
-  EXPIRED
-  ;
-}
