--- service/cart/src/main/java/com/bappul/delivery/cart/domain/repository/CartItemOptionRepository.java
@@ -1,12 +1,0 @@
-package com.bappul.delivery.cart.domain.repository;
-
-import com.bappul.delivery.cart.domain.entity.CartItem;
-import com.bappul.delivery.cart.domain.entity.CartItemOption;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CartItemOptionRepository extends JpaRepository<CartItemOption, Long> {
-  List<CartItemOption> findAllByCartItem(CartItem cartItem);
-}
