--- service/cart/src/main/java/com/bappul/delivery/cart/web/v1/request/CartItemRequest.java
@@ -1,11 +1,0 @@
-package com.bappul.delivery.cart.web.v1.request;
-
-import java.util.List;
-import lombok.Getter;
-
-@Getter
-public class CartItemRequest {
-  Long menuId;
-  List<Long> optionValueIds;
-  int quantity;
-}
